Update descendant search to use first-match instead of deepest match
This should fix the sibling index miscalculation when figuring out where to insert a new element. In general, it finds the nearest parent that has a physical control, and then based on where the current adapter is, it finds the previous sibling that also has a physical control in the aforementioned parent. Hopefully, it works.