Fixed Error when using CSS on iOS
Updated StyleSheetHandler to use a diferent Xamarin Forms method for loading the CSS File.

When it was using FromResource we would get a NullReferenceException when Forms tries to find an IResourceLoader.

This fix uses an obsolete method that suggests using the version that was already in use which is concerning. Perhaps the real fix needs to happen in the Forms repo.