Added placeholder page on launch and update ElementManager to allow replacing MainPage
This fixes issue #138, but is a bit of a workaround rather than a proper fix, but adding this now doesn't cause any real harm or prevent more work on it later.

I've put more details about why I couldn't fix the async call in the issue comments.

# Changes
Added a placeholder page in the App constructor so that there is a root view strain away.

Updated MobileBlazorBindingsElementManager to allow MainPages being swapped at run time, there was previously a check to make sure MainPage wasn't already set, which I don't think is necessary.

Changes in MobileBlazorBindingsElementManager  look a lot more complex than they are because indentation changed when I remove the if check.