Implemented Type Safety and Parameter checking in Navigation
Created a converter to handle the 9 types of valid navigaiton parameteres, we now try to convert everything to the appropriate type bfore setting the property.

Also added a check fo the [Parameter] attribute so we will only set properties marked as paramerters. I'm not convinced this is a good idea but it seems to be how it's done in Blazor. There's probably a good reason behind it.