Native Component Code Review Changes
Changed the Invalid Cast Exception to ad Invalid Operation Excption to better reflect the error when a user tries to use an unsupported type.

Changed the default result value for unsupported types. It was a 5, but now it's a null. The 5 never should have been there.

Unit tests updated for both changes.